Fix scrolling with new scrollSection component
