+++ src/components/FeaturedMovie/FeaturedMovie.css
@@ -38,7 +38,7 @@
   flex-direction: column;
   align-items: flex-start;
   justify-content: flex-end;
-  z-index: 10;
+  z-index: 5;
 }
