+++ src/components/ScrollMovies/ScrollMovies.js
@@ -2,22 +2,10 @@ import React from 'react';
 import './ScrollMovies.css';
 import Card from '../Card/Card';
 import FeaturedMovie from '../FeaturedMovie/FeaturedMovie'
-import ArrowBackIosIcon from '@material-ui/icons/ArrowBackIos';
-import ArrowForwardIosIcon from '@material-ui/icons/ArrowForwardIos';
-
+import ScrollSection from '../ScrollSection/ScrollSection';
 
 
 const ScrollMovies = ({movies, allGenres}) => {
-  const scroller = React.useRef()
-
-  const scrollRight = () => {
-    scroller.current.scrollLeft += 200
-  }
-
-  const scrollLeft = () => {
-    scroller.current.scrollLeft -= 200
-  }
-
   const randomMovie = () => {
     const randomIndex = Math.floor(Math.random() * Math.floor(movies.length))
     return movies[randomIndex]
@@ -29,36 +17,12 @@ const ScrollMovies = ({movies, allGenres}) => {
     <div className="movies-genres">
       {allGenres.map(genre => {
       return (
-        <div className="genre-row">
-          <div className= "genre">
-            <h2 className="genre-name">{`${genre}`}</h2>
-            <div className="slider">
-              <span 
-                className="leftScroll"  
-                tabIndex="0" 
-                role="button"   
-                aria-label="See previous titles"
-                onClick={scrollLeft}
-              >
-                <ArrowBackIosIcon />
-              </span>
-                <div className="slider-cards" ref={scroller}>
-                  {movies
-                    .filter(movie => movie.genres.includes(genre))
-                    .map(movie => <Card movie={movie} />)
-                  }
-                </div>
-              <span 
-                className="rightScroll"  
-                tabIndex="0" role="button"   
-                aria-label="See previous titles" 
-                onClick={scrollRight}
-              >
-                <ArrowForwardIosIcon />
-              </span>
-            </div>
-          </div>
-        </div>
+       <ScrollSection 
+          movies={movies}
+          allGenres={allGenres}
+          genre={genre}
+          key={`scroll-section-${genre}`}
+        />
       )
     })}
     </div>
